# desc/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Fishing Notes Admin API",
    "version": "0.2.1",
    "description": "管理后台 API 文档，覆盖认证、RBAC、管理员用户、钓点审核、知识专栏、鱼种内容与业务用户管理。\n\n使用说明：\n1. 除登录接口外，其余接口默认都需要 Bearer Token。\n2. 所有成功响应统一为 `{ code, message, data }`，其中 `code = 0` 表示成功。\n3. 列表接口统一支持 `page` 和 `pageSize`，默认 `page=1`、`pageSize=20`。\n"
  },
  "servers": [
    {
      "url": "http://localhost:1002",
      "description": "Local API Server"
    }
  ],
...
  "security": [
    {
      "bearerAuth": [

      ]
    }
  ],
  "paths": {
...
    "/api/v1/review-tasks": {
      "get": {
        "tags": [
          "Review Tasks"
        ],
        "summary": "获取统一审核任务",
        "operationId": "listReviewTasks",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "keyword",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected"
              ]
            }
          },
          {
            "name": "taskType",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "spot",
                "user_spot",
                "spot_correction",
                "public_report"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListReviewTasksEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/audit-logs": {
      "get": {
        "tags": [
          "Review Tasks"
        ],
        "summary": "查询审核历史",
        "operationId": "listAuditLogs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          },
          {
            "name": "resourceType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resourceId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListAuditLogsEnvelope"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/audit-logs/export": {
      "get": {
        "tags": [
          "Review Tasks"
        ],
        "summary": "导出操作日志",
        "operationId": "exportAuditLogs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "resourceType",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resourceId",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "action",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CSV 文件，最多导出最近 10000 条日志",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {
              "X-Export-Truncated": {
                "description": "是否因 10000 条上限截断",
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/v1/review-tasks/public-reports/{recordId}/approve": {
      "post": {
        "tags": [
          "Review Tasks"
        ],
        "summary": "通过公开鱼情审核",
        "operationId": "approvePublicFishingReport",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RecordIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewPublicFishingReportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
    "/api/v1/review-tasks/public-reports/{recordId}/reject": {
      "post": {
        "tags": [
          "Review Tasks"
        ],
        "summary": "驳回公开鱼情审核",
        "operationId": "rejectPublicFishingReport",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/RecordIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewPublicFishingReportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "操作成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/InternalServerError" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "parameters": {
      "Page": {
        "name": "page",
        "in": "query",
        "description": "页码，默认 1",
        "schema": {
          "type": "integer",
          "format": "int64",
          "default": 1
        }
      },
      "PageSize": {
        "name": "pageSize",
        "in": "query",
        "description": "每页数量，默认 20，最大 100",
        "schema": {
          "type": "integer",
          "format": "int64",
          "default": 20,
          "maximum": 100
        }
      },
...
      "RecordIdPath": {
        "name": "recordId",
        "in": "path",
        "required": true,
        "description": "渔获记录 ID",
        "schema": {
          "type": "integer",
          "format": "int64"
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "请求参数错误",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "未认证或认证失败",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "没有权限访问",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "NotFound": {
        "description": "资源不存在",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "服务端异常",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    },
    "schemas": {
...
      "ReviewPublicFishingReportRequest": {
        "type": "object",
        "required": [
          "reviewNote"
        ],
        "properties": {
          "reviewNote": {
            "type": "string",
            "maxLength": 500
          }
        }
      },
      "ReviewTaskItem": {
        "type": "object",
        "properties": {
          "taskId": {
            "type": "string"
          },
          "taskType": {
            "type": "string",
            "enum": [
              "spot",
              "user_spot",
              "spot_correction",
              "public_report"
            ]
          },
          "resourceId": {
            "type": "integer",
            "format": "int64"
          },
          "resourceCode": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "approved",
              "rejected"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "AuditLogItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "actorUserId": {
            "type": "integer",
            "format": "int64"
          },
          "action": {
            "type": "string"
          },
          "resourceType": {
            "type": "string"
          },
          "resourceId": {
            "type": "integer",
            "format": "int64"
          },
          "detail": {
            "type": "object",
            "additionalProperties": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ListAuditLogsResponse": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AuditLogItem"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int64"
          },
          "pageSize": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ListReviewTasksResponse": {
        "type": "object",
        "properties": {
          "list": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReviewTaskItem"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int64"
          },
          "pageSize": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "SuccessData": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "example": {
          "success": true
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "example": {
          "code": 10000,
          "message": "invalid request"
        }
      },
...
      "ListReviewTasksEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/ListReviewTasksResponse"
          }
        }
      },
      "ListAuditLogsEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/ListAuditLogsResponse"
          }
        }
      },
...
      "SuccessEnvelope": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "example": 0
          },
          "message": {
            "type": "string",
            "example": "ok"
          },
          "data": {
            "$ref": "#/components/schemas/SuccessData"
          }
        },
        "example": {
          "code": 0,
          "message": "ok",
          "data": {
            "success": true
          }
        }
      }
    }
  }
}